backend/requirements.txt
Django>=4.2,<5.0
djangorestframework>=3.14
djangorestframework-simplejwt>=5.3
django-cors-headers>=4.3
django-filter>=23.5
cloudinary>=1.36
python-decouple>=3.8
psycopg2-binary>=2.9
Pillow>=10.1
drf-spectacular>=0.27
python-dotenv>=1.0
django-cloudinary-storage>=0.3.0
